fix(prefix-tab): cast $8 to BIGINT so tokio-postgres serializes Option<i64> correctly
The INSERT for generation_ms used $8 with no type hint. tokio-postgres
cannot infer the wire type for Option<i64> without a cast annotation,
causing 'error serializing parameter 4' (0-indexed) and rejecting every
store() call with a DB error. Adding ::BIGINT gives tokio-postgres the
explicit OID it needs.

This was the root cause of 'Regenerate failed: unavailable' on every
Generate click — the assembly completed successfully but store() always
failed, returning 503.

+++ crates/brassclaw_reborn_composition/src/pg_basic_prompt_store.rs
@@ -204,7 +204,7 @@ mod inner {
                           assembled_at, prewarm_last_at, updated_at, generation_ms)
                      VALUES ($1, $2, $3, $4, $5::JSONB, $6, false, now(),
                              CASE WHEN $7 THEN now() ELSE NULL END,
-                             now(), $8)
+                             now(), $8::BIGINT)
                      ON CONFLICT ON CONSTRAINT reborn_basic_prompt_store_scope_unique
                      DO UPDATE SET
                          bundle_json     = EXCLUDED.bundle_json,
